set dtype=64 to fix dtype problem in windows

+++ applications/sentiment_analysis/deploy/predict.py
@@ -109,9 +109,9 @@ class Predictor(object):
         ]
 
         batchify_fn = lambda samples, fn=Tuple(
-            Pad(axis=0, pad_val=self.tokenizer.pad_token_id),
-            Pad(axis=0, pad_val=self.tokenizer.pad_token_type_id),
-            Stack(dtype="int64"), ): fn(samples)
+            Pad(axis=0, pad_val=self.tokenizer.pad_token_id, dtype="int64"),
+            Pad(axis=0, pad_val=self.tokenizer.pad_token_type_id, dtype="int64"),
+            Stack(dtype="int64")): fn(samples)
 
         results = []
         for bid, batch_data in enumerate(batch_list):
@@ -158,8 +158,8 @@ class Predictor(object):
         ]
 
         batchify_fn = lambda samples, fn=Tuple(
-            Pad(axis=0, pad_val=self.tokenizer.pad_token_id),
-            Pad(axis=0, pad_val=self.tokenizer.pad_token_type_id),
+            Pad(axis=0, pad_val=self.tokenizer.pad_token_id, dtype="int64"),
+            Pad(axis=0, pad_val=self.tokenizer.pad_token_type_id, dtype="int64"),
             Stack(dtype="int64")): fn(samples)
 
         results = []
